Moved App.js to components
