+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './components/App';
 import { createStore } from 'redux'
 import { Provider } from 'react-redux'
 import reducers from '../src/reducers'
